Move backend api config to existing config folder.
