+++ server/backend/api/app/database.py
@@ -5,7 +5,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from . import config
+from api.config import config
 
 LOGGER = logging.getLogger(__name__)
